+++ src/app/products/models/category.enum.ts
@@ -1,0 +1,9 @@
+export enum Category {
+  Clothes = 'clothes',
+  Electronics = 'electronics',
+  Food = 'food',
+  Furniture = 'furniture',
+  Health = 'health',
+  KitchenUtensils = 'kitchen utensils',
+  Toys = 'toys'
+}