+++ src/app/products/models/index.ts
@@ -1,0 +1,3 @@
+export * from './category.enum';
+export * from './payment.enum';
+export * from './product.model';