+++ src/app/shared/directives/paint-background.directive.spec.ts
@@ -1,0 +1,8 @@
+import { PaintBackgroundDirective } from './paint-background.directive';
+
+describe('PaintBackgroundDirective', () => {
+  it('should create an instance', () => {
+    const directive = new PaintBackgroundDirective();
+    expect(directive).toBeTruthy();
+  });
+});